requirements.txt
transformers==4.41.2
scikit-learn==1.5.0
pdfplumber==0.10.4
python-docx==1.1.0